define downloder in subcommand script
