+++ gl2f/core/article.py
@@ -143,84 +143,6 @@ def dl_medium(boardId, contentId, mediaId, head=False, stream=False, streamfile=
 	else:
 		return meta, requests.get(url, stream=stream)
 
-
-class Bar:
-	def __init__(self, li, contentId):
-		from threading import Lock
-
-		self.n = len(li)
-		self.dig = len(str(self.n))
-
-		w, _ = os.get_terminal_size()
-		self.width = w - 2*self.dig - 26
-
-		self.lock = Lock()
-
-		self.contentId = contentId
-
-		self.progress = {k:{'progress':0, 'length':1} for k in li}
-
-
-	def bar(self):
-		f = sum(p['progress']/p['length'] for p in self.progress.values()) / len(self.progress)
-		i = int(f*self.width)
-		return f'[{"#"*i}{"-"*(self.width-i)}]'
-
-	def count(self):
-		i = sum(1 for _ in (i['progress'] for i in self.progress.values() if i['progress']>0))
-		return f'[{i:{self.dig}}/{self.n:{self.dig}}]'
-
-	def print(self):
-		with self.lock:
-			term.clean_row()
-			print(f'{self.contentId} {self.bar()} {self.count()}', end='', flush=True)
-
-
-def save_media(item, out, boardId, contentId,
-	skip=False, streamfile=False, force=False, dump=False
-):
-	from concurrent.futures import ThreadPoolExecutor
-	from functools import partial
-
-	li = [i.group(1) for i in ptn_media.finditer(item['values']['body'])]
-
-	bar = Bar(li, contentId)
-	bar.print()
-
-	xauth = auth.update(auth.load())
-	_dl = partial(dl_medium, boardId, contentId, head=skip, stream=True, streamfile=streamfile, xauth=xauth)
-
-	def dl(mediaId):
-		ptn = re.compile(mediaId + r'\..+')
-		if (not force) and any(map(ptn.search, os.listdir(out))):
-			return 'skipped'
-
-		meta, response = _dl(mediaId=mediaId)
-
-		bar.progress[mediaId]['length'] = int(response.headers['content-length'])
-
-		with open(os.path.join(out, f'{meta["mediaId"]}.{meta["meta"]["ext"]}'), 'wb') as f:
-			for i in response.iter_content(chunk_size=1024*1024):
-				f.write(i)
-
-				bar.progress[mediaId]['progress'] += len(i)
-				bar.print()
-
-
-		response.close()
-
-		return meta
-
-
-	with ThreadPoolExecutor() as executor:
-		futures = [executor.submit(dl, i) for i in li]
-
-	if dump:
-		now = datetime.datetime.now().strftime('%y%m%d%H%M%S')
-		with open(os.path.join(dump, f'media-{contentId}-{now}.json'), 'w', encoding='utf-8') as f:
-			json.dump([f.result() for f in futures], f, indent=2, ensure_ascii=False)
-
-
 def media_stat(body):
 	types = [t for _, t in  ptn_media.findall(body)]
 	return {key: types.count(key) for key in ['image', 'video']}
